+++ client/src/components/admin/Events.tsx
@@ -14,7 +14,13 @@ import {
   SelectInput,
   ReferenceField,
   ReferenceInput,
-  AutocompleteInput
+  AutocompleteInput,
+  ReferenceManyField,
+  ReferenceArrayInput,
+  AutocompleteArrayInput,
+  ArrayField,
+  SingleFieldList,
+  ChipField
 } from 'react-admin'
 
 export const EventList = () => (
@@ -24,9 +30,15 @@ export const EventList = () => (
       <TextField source="title" />
       <TextField source="type" />
       <NumberField source="chapterNumber" />
+      <NumberField source="spoilerChapter" label="Spoiler Ch." />
       <ReferenceField source="arcId" reference="arcs" label="Arc">
         <TextField source="name" />
       </ReferenceField>
+      <ArrayField source="characters" label="Characters">
+        <SingleFieldList linkType={false}>
+          <ChipField source="name" size="small" />
+        </SingleFieldList>
+      </ArrayField>
       <TextField source="description" />
     </Datagrid>
   </List>
@@ -44,6 +56,11 @@ export const EventShow = () => (
       <ReferenceField source="arcId" reference="arcs" label="Arc">
         <TextField source="name" />
       </ReferenceField>
+      <ArrayField source="characters" label="Characters">
+        <SingleFieldList linkType={false}>
+          <ChipField source="name" />
+        </SingleFieldList>
+      </ArrayField>
     </SimpleShowLayout>
   </Show>
 )
@@ -66,11 +83,14 @@ export const EventEdit = () => (
         ]}
         required
       />
-      <NumberInput source="chapterNumber" required max={539} />
-      <NumberInput source="spoilerChapter" max={539} />
+      <NumberInput source="chapterNumber" required max={539} min={1} />
+      <NumberInput source="spoilerChapter" max={539} min={1} />
       <ReferenceInput source="arcId" reference="arcs" label="Arc">
         <AutocompleteInput optionText="name" />
       </ReferenceInput>
+      <ReferenceArrayInput source="characterIds" reference="characters" label="Characters">
+        <AutocompleteArrayInput optionText="name" />
+      </ReferenceArrayInput>
     </SimpleForm>
   </Edit>
 )
@@ -94,11 +114,14 @@ export const EventCreate = () => (
         required
         defaultValue="other"
       />
-      <NumberInput source="chapterNumber" required max={539} />
-      <NumberInput source="spoilerChapter" max={539} />
+      <NumberInput source="chapterNumber" required max={539} min={1} />
+      <NumberInput source="spoilerChapter" max={539} min={1} />
       <ReferenceInput source="arcId" reference="arcs" label="Arc">
         <AutocompleteInput optionText="name" />
       </ReferenceInput>
+      <ReferenceArrayInput source="characterIds" reference="characters" label="Characters">
+        <AutocompleteArrayInput optionText="name" />
+      </ReferenceArrayInput>
     </SimpleForm>
   </Create>
 )